Fix numeric overflows

+++ GithubAnalytics/Json/PullRequestFile.cs
@@ -4,9 +4,9 @@ public class PullRequestFile
 {
 	public string Filename { get; set; }
 
-	public int Additions { get; set; }
-	public int Deletions { get; set; }
-	public int Changes { get; set; }
+	public long Additions { get; set; }
+	public long Deletions { get; set; }
+	public long Changes { get; set; }
 
 	//  added | modified | ???
 	public string Status { get; set; }
